Modified architecture to be able to read fonction
The fb_send function now passes the message from the dialog.json through the FilterTags.
FilterTags searches in the message for %%FUNCTION%% Tags then executes this function and replace the tags and call by the returned value of the so called function.

On the other end, MrYoung.js checks for tags in the messages to search more easily in the DB for special messages that inlfuence a background function